elections/requirements.txt
pymysql==1.0.2
Flask==1.1.2
flask_sqlalchemy==2.5.1
sqlalchemy <1.4.0
flask_migrate==2.7.0
flask_script==2.0.6
sqlalchemy_utils==0.36.8
flask_jwt_extended==4.1.0
cryptography==3.4.7
redis==3.5.3
python-dateutil==2.8.2


